Diagram of data architechture
this architecture is a very simplified view of some basic pieces for a cold & hot (ETL and streamed )  pipeline to availability. all pieces are available as serverless, so that availability is served, the internal secrets manager shows a division of concearns from other segments of the business ecosistem, so that data and application personnel might keep to themselves, and greater governance is stablished in these two segments. a firewall is present at the VPN, such that there is a path to Friend Or Foe capabilities is available, enforcing security. the cold trail also defines greater complience, such that data labeled as efemeral might be simply discarded from kiness retention, never entering any write now delete whenever cycles.